+++ src/workload-orchestration/azext_workload_orchestration/aaz/latest/workload_orchestration/context/_show.py
@@ -61,6 +61,7 @@ class Show(AAZCommand):
 
     def _execute_operations(self):
         self.pre_operations()
+
         self.ContextsGet(ctx=self.ctx)()
         self.post_operations()