+++ src/confcom/HISTORY.rst
@@ -3,6 +3,13 @@
 Release History
 ===============
 
+1.2.7
+++++++
+* bugfix making it so that oras discover function doesn't error when no fragments are found in the remote repository
+* splitting out documentation into command-specific files and adding info about --input flag
+* adding standalone fragment support
+* bugfix for oras pulling fragments when offline
+
 1.2.6
 ++++++
 * bugfix making it so the fields in the --input format are case-insensitive